3-0-java-core/3-6-1-file-reader/src/test/java/com/bobocode/se/FileReadersTest.java
package com.bobocode.se;

import org.junit.jupiter.api.Test;

import java.io.IOException;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class FileReadersTest {

    @Test
    void testReadWholeFileOnEmptyFile() throws IOException {
        String fileContent = FileReaders.readWholeFile("empty.txt");

        assertEquals("", fileContent);

    }

    @Test
    void testReadWholeFileOnFileWithEmptyLines() throws IOException {
        String fileContent = FileReaders.readWholeFile("lines.txt");

        assertEquals("Hey!\n" +
                "\n" +
                "What's up?\n" +
                "\n" +
                "Hi!", fileContent);
    }

    @Test
    void testReadWholeFile() throws IOException {
        String fileContent = FileReaders.readWholeFile("simple.txt");

        assertEquals("Hello!\n" + "It's a test file.", fileContent);
    }
}
